+++ verik-plugin/src/main/kotlin/io/verik/plugin/domain/DsimTargetDomainObject.kt
@@ -7,9 +7,9 @@ package io.verik.plugin.domain
 import java.nio.file.Path
 
 /**
- * Interface for domain objects that configure an aurora target.
+ * Interface for domain objects that configure a Metrics dsim target.
  */
-interface AuroraTargetDomainObject : TargetDomainObject {
+interface DsimTargetDomainObject : TargetDomainObject {
 
     var compileTops: List<String>
     var dpiLibs: List<Path>