+++ verik-plugin/src/main/kotlin/io/verik/plugin/domain/DsimTargetDomainObjectImpl.kt
@@ -7,11 +7,11 @@ package io.verik.plugin.domain
 import java.nio.file.Path
 
 /**
- * Implementation for domain objects that configure an aurora target.
+ * Implementation for domain objects that configure a metrics Dsim target.
  */
-class AuroraTargetDomainObjectImpl : AuroraTargetDomainObject {
+class DsimTargetDomainObjectImpl : DsimTargetDomainObject {
 
-    override var name: String = "aurora"
+    override var name: String = "dsim"
     override var compileTops: List<String> = listOf()
     override var dpiLibs: List<Path> = listOf()
 }