+++ verik-plugin/src/main/kotlin/io/verik/plugin/target/DsimTargetBuilder.kt
@@ -1,0 +1,15 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.verik.plugin.target
+
+import io.verik.plugin.config.DsimTargetConfig
+
+/**
+ * Builder for Metrics dsim targets.
+ */
+object DsimTargetBuilder {
+
+    fun build(targetConfig: DsimTargetConfig) {}
+}