+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - Move compiler to `verikCompile` task.
 - Add `verik` task that generates build scripts for downstream tools.
 - Move compiler and importer build directories to `build/compile` and `build/import`.
+- Add the Metrics dsim generation target.
+- Add the Xilinx vivado generation target.
 
 ## [0.1.15]
 ### Compiler